refactor: migrate post comment functionality to commenting package
