+++ src/main/java/org/nfactorial/newsfeed/domain/commenting/controller/PostCommentController.java
@@ -1,16 +1,16 @@
-package org.nfactorial.newsfeed.domain.postcomment.controller;
+package org.nfactorial.newsfeed.domain.commenting.controller;
 
 import org.nfactorial.newsfeed.common.code.SuccessCode;
 import org.nfactorial.newsfeed.common.dto.GlobalApiResponse;
 import org.nfactorial.newsfeed.common.security.AuthProfile;
 import org.nfactorial.newsfeed.common.security.AuthProfileDto;
-import org.nfactorial.newsfeed.domain.postcomment.dto.command.WriteCommentToPostCommand;
-import org.nfactorial.newsfeed.domain.postcomment.dto.request.WriteCommentToPostRequest;
-import org.nfactorial.newsfeed.domain.postcomment.dto.response.GetCommentsFromPostResponse;
-import org.nfactorial.newsfeed.domain.postcomment.dto.response.WriteCommentToPostResponse;
-import org.nfactorial.newsfeed.domain.postcomment.dto.result.CommentListByPostResult;
-import org.nfactorial.newsfeed.domain.postcomment.dto.result.WriteCommentToPostResult;
-import org.nfactorial.newsfeed.domain.postcomment.service.PostCommentService;
+import org.nfactorial.newsfeed.domain.commenting.dto.command.WriteCommentToPostCommand;
+import org.nfactorial.newsfeed.domain.commenting.dto.request.WriteCommentToPostRequest;
+import org.nfactorial.newsfeed.domain.commenting.dto.response.GetCommentsFromPostResponse;
+import org.nfactorial.newsfeed.domain.commenting.dto.response.WriteCommentToPostResponse;
+import org.nfactorial.newsfeed.domain.commenting.dto.result.CommentListByPostResult;
+import org.nfactorial.newsfeed.domain.commenting.dto.result.WriteCommentToPostResult;
+import org.nfactorial.newsfeed.domain.commenting.service.PostCommentService;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
