+++ src/main/java/org/nfactorial/newsfeed/domain/commenting/dto/response/GetCommentsFromPostResponse.java
@@ -1,8 +1,8 @@
-package org.nfactorial.newsfeed.domain.postcomment.dto.response;
+package org.nfactorial.newsfeed.domain.commenting.dto.response;
 
 import java.util.List;
 
-import org.nfactorial.newsfeed.domain.postcomment.dto.result.CommentListByPostResult;
+import org.nfactorial.newsfeed.domain.commenting.dto.result.CommentListByPostResult;
 
 public record GetCommentsFromPostResponse(
 	List<SimpleComment> comments) {
