+++ src/test/java/org/nfactorial/newsfeed/domain/comment/WriteCommentE2ETest.java
@@ -7,8 +7,8 @@ import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.nfactorial.newsfeed.common.dto.GlobalApiResponse;
-import org.nfactorial.newsfeed.domain.postcomment.dto.request.WriteCommentToPostRequest;
-import org.nfactorial.newsfeed.domain.postcomment.dto.response.WriteCommentToPostResponse;
+import org.nfactorial.newsfeed.domain.commenting.dto.request.WriteCommentToPostRequest;
+import org.nfactorial.newsfeed.domain.commenting.dto.response.WriteCommentToPostResponse;
 import org.springframework.core.ParameterizedTypeReference;
 import org.springframework.http.HttpEntity;
 import org.springframework.http.HttpMethod;
